requirements.txt
django==2.0
scapy-python3
peewee